backend/requirements.txt
fastapi
uvicorn
nest-asyncio
pyngrok
python-multipart
pydantic
sentence-transformers
chromadb
groq
langchain
langchain-community
langchain-groq
python-dotenv
